+++ packages/iron-realms-nexus-typings/definitions/index.d.ts
@@ -1,3 +1,3 @@
-export * from './client-api'
-export * from './Reflexes/index'
-export * from './Nexus3/index'
+export * from './client-api';
+export * from './Reflexes/index';
+export * from './Nexus3/index';