+++ packages/nexus-package-builder/src/classes/keybind.ts
@@ -1,5 +1,6 @@
 import { PartialKeybind } from '../types';
 import { convertNexusActionArray, IdGenerator } from '../utils';
+import * as client from '@keneanung/iron-realms-nexus-typings';
 
 /**
  * Class for Nexus keybind