+++ packages/nexus-package-builder/src/classes/notifyAction.ts
@@ -1,3 +1,4 @@
+import * as client from '@keneanung/iron-realms-nexus-typings';
 /**
  * Class for a Nexus notify action.
  */