+++ packages/nexus-event-bus/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.9](https://github.com/keneanung/nexus-scripts/compare/@keneanung/nexus-event-bus@0.5.8...@keneanung/nexus-event-bus@0.5.9) (2023-07-17)
+
+**Note:** Version bump only for package @keneanung/nexus-event-bus
+
 ## [0.5.8](https://github.com/keneanung/nexus-scripts/compare/@keneanung/nexus-event-bus@0.5.7...@keneanung/nexus-event-bus@0.5.8) (2023-07-04)
 
 **Note:** Version bump only for package @keneanung/nexus-event-bus