+++ packages/nexus-package-builder/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.7](https://github.com/keneanung/nexus-package-builder/compare/@keneanung/nexus-package-builder@1.3.6...@keneanung/nexus-package-builder@1.3.7) (2023-07-17)
+
+### Bug Fixes
+
+- **nexus-package-builder:** :arrow_up: use latest version of nexus typings ([8e233ab](https://github.com/keneanung/nexus-package-builder/commit/8e233abdf02e4390602b272519b654db220a41b5))
+
 ## [1.3.6](https://github.com/keneanung/nexus-package-builder/compare/@keneanung/nexus-package-builder@1.3.5...@keneanung/nexus-package-builder@1.3.6) (2023-07-04)
 
 **Note:** Version bump only for package @keneanung/nexus-package-builder