+++ packages/nexus-package-manager/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/keneanung/nexus-package-manager/compare/@keneanung/nexus-package-manager@0.3.1...@keneanung/nexus-package-manager@0.3.2) (2023-07-17)
+
+### Bug Fixes
+
+- **nexus-package-manager:** add changes required by typing fixes ([c24e6eb](https://github.com/keneanung/nexus-package-manager/commit/c24e6eb5bd15c0a922def7773c1456d9a645a3b6))
+
 ## [0.3.1](https://github.com/keneanung/nexus-package-manager/compare/@keneanung/nexus-package-manager@0.3.0...@keneanung/nexus-package-manager@0.3.1) (2023-07-04)
 
 ### Bug Fixes