+++ packages/nexus-queue-manager/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2](https://github.com/keneanung/nexus-scripts/compare/@keneanung/nexus-queue-manager@0.1.1...@keneanung/nexus-queue-manager@0.1.2) (2023-07-17)
+
+**Note:** Version bump only for package @keneanung/nexus-queue-manager
+
 ## [0.1.1](https://github.com/keneanung/nexus-scripts/compare/@keneanung/nexus-queue-manager@0.1.0...@keneanung/nexus-queue-manager@0.1.1) (2023-07-04)
 
 **Note:** Version bump only for package @keneanung/nexus-queue-manager